test tecnicoMasCercano funca

+++ src/main/java/ar/edu/utn/frba/dds/models/puntoIdeal/IAdapterPuntoIdeal.java
@@ -4,5 +4,5 @@ import ar.edu.utn.frba.dds.models.data.Ubicacion;
 import java.util.List;
 
 public interface IAdapterPuntoIdeal {
-  List<Ubicacion> puntoIdeal(Integer latitud, Integer longitud, Integer radio);
+  List<Ubicacion> puntoIdeal(Double latitud, Double longitud, Double radio);
 }
